Extract llm-logger from pipeline/ to observability/
The pipeline/ directory is offline tooling -- the policy compiler. Runtime
modules (sandbox, session) were importing the LLM logging middleware factory
from there, which is a layer violation in the same vein as the recent
getListMatcher extraction (#219).

Pure file move plus import path updates. No API or behavior changes.

The new src/observability/ directory makes room for future structured-runtime-
data utilities (metric counters, audit-stream tailers, etc.) without crowding
the top level or pipeline/.
